Move TypedBounds into typed_bounds module
