+++ src/lib.rs
@@ -10,6 +10,7 @@ pub mod tuple_index;
 pub mod tuple_key;
 pub mod typed_first;
 pub mod typed_index;
+pub mod typed_bounds;
 pub mod typed_last;
 pub mod typed_nth;
 pub mod typed_tuple;
@@ -26,7 +27,8 @@ pub mod prelude {
         tuple_index::TupleIndex,
         tuple_key::TupleKey,
         typed_first::TypedFirst,
-        typed_index::{TypedIndex, TypedBounds},
+        typed_index::TypedIndex,
+        typed_bounds::TypedBounds,
         typed_last::TypedLast,
         typed_nth::TypedNth,
         typed_tuple::TypedTuple,
